examples refactor
