--- examples/README.md
@@ -1,12 +1,0 @@
-# Examples
-This directory contains examples of how to use the egui_graphs library.
-
-## Prerequisites
-All examples require rust to be installed. Just follow easy one step install from the official site https://www.rust-lang.org/tools/install.
-
-## Run
-To run an example, use the following command:
-```bash
-cargo run --release -p <EXAMPLE_NAME>
-```
-where `<EXAMPLE_NAME>` is the name of the example you want to run.
