--- examples/basic/README.md
@@ -1,9 +1,0 @@
-# Basic
-
-Basic example which demonstrates the usage of `GraphView` widget.
-
-## run
-
-```bash
-cargo run --release -p basic
-```
