--- examples/basic_custom/README.md
@@ -1,9 +1,0 @@
-# Basic
-
-Basic example which demonstrates the usage of `GraphView` widget with helper functions.
-
-## run
-
-```bash
-cargo run --release -p basic_custom
-```
