--- examples/bevy_basic/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "bevy_basic"
-version = "0.1.0"
-authors = ["Dmitrii Samsonov <blitzarx1@gmail.com>"]
-license = "MIT"
-edition = "2021"
-
-[dependencies]
-egui_graphs = { path = "../.." }
-bevy = "0.15"
-bevy_egui = "0.33"
-petgraph = "0.8"
