--- examples/bevy_basic/README.md
@@ -1,7 +1,0 @@
-# Bevy Basic
-Basic example which demonstrates the usage of `GraphView` widget in Bevy.
-
-## run
-```bash
-cargo run --release -p bevy_basic
-```
