--- examples/flex_nodes/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "flex_nodes"
-version = "0.1.0"
-authors = ["Dmitrii Samsonov <blitzarx1@gmail.com>"]
-license = "MIT"
-edition = "2021"
-
-[dependencies]
-egui_graphs = { path = "../.." }
-egui = "0.31"
-eframe = "0.31"
-petgraph = "0.8"
