--- examples/flex_nodes/README.md
@@ -1,9 +1,0 @@
-# Flex nodes
-
-Example demonstrates how to use custom drawing functions for nodes with flexible size which depends on their label size. 
-
-## run
-
-```bash
-cargo run --release -p flex_nodes
-```
