--- examples/interactive/README.md
@@ -1,11 +1,0 @@
-# Interactive
-
-Modification of the [basic example](https://github.com/blitzarx1/egui_graph/tree/master/examples/basic) which shows how easy it is to enable interactivity.
-
-Try to drag around and select some nodes.
-
-## run
-
-```bash
-cargo run --release -p interactive
-```
