--- examples/label_change/README.md
@@ -1,9 +1,0 @@
-# Label Change
-
-Simple example which demonstrates the API for changing node props including specifically `label`.
-
-## run
-
-```bash
-cargo run --release -p label_change
-```
