--- examples/layouts/README.md
@@ -1,9 +1,0 @@
-# Layouts
-
-Example of usage of different lyouts implemented by `egui_graphs`.
-
-## run
-
-```bash
-cargo run --release -p layouts
-```
