--- examples/multiple/README.md
@@ -1,9 +1,0 @@
-# Multiple
-
-Example which demonstrates the usage of `GraphView` widget in multiple egui tabs. Please pay attention that there is one graph and multiple `GraphView` widgets - this means that modifying graph in one widget causes modification in others.
-
-## run
-
-```bash
-cargo run --release -p multiple
-```
