--- examples/rainbow_edges/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "rainbow_edges"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-egui_graphs = { path = "../.." }
-egui = "0.31"
-eframe = "0.31"
-petgraph = "0.8"
