--- examples/rainbow_edges/README.md
@@ -1,9 +1,0 @@
-# Rainbow Edges
-
-Example demonstrates the possibility of drawing custom edges shapes.
-
-## run
-
-```bash
-cargo run --release -p rainbow_edges
-```
