--- examples/undirected/README.md
@@ -1,9 +1,0 @@
-# Undirected
-
-Basic example which demonstrates the usage of `GraphView` widget with undirected graph.
-
-## run
-
-```bash
-cargo run --release -p undirected
-```
