--- examples/wasm_custom_draw/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "wasm_custom_draw"
-version = "0.1.0"
-authors = ["Atlas16A"]
-license = "MIT"
-edition = "2021"
-
-[dependencies]
-egui_graphs = { path = "../.." }
-egui = "0.31"
-eframe = "0.31"
-petgraph = "0.8"
-
-# Wasm related dependencies
-getrandom = { version = "0.2", features = ["js"] }
-log = "0.4"
-instant = { version = "0.1", features = ["wasm-bindgen"] }
-
-
-# native:
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-env_logger = "0.11"
-
-# web:
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-wasm-bindgen-futures = "0.4"
-
-
-[profile.release]
-opt-level = 2 # fast and small wasm
