--- examples/wasm_custom_draw/Trunk.toml
@@ -1,1 +1,0 @@
-[build]
