--- examples/wasm_custom_draw/assets/manifest.json
@@ -1,28 +1,0 @@
-{
-  "name": "wasm",
-  "short_name": "wasm",
-  "icons": [
-    {
-      "src": "./icon-256.png",
-      "sizes": "256x256",
-      "type": "image/png"
-    },
-    {
-      "src": "./maskable_icon_x512.png",
-      "sizes": "512x512",
-      "type": "image/png",
-      "purpose": "any maskable"
-    },
-    {
-      "src": "./icon-1024.png",
-      "sizes": "1024x1024",
-      "type": "image/png"
-    }
-  ],
-  "lang": "en-US",
-  "id": "/index.html",
-  "start_url": "./index.html",
-  "display": "standalone",
-  "background_color": "white",
-  "theme_color": "white"
-}
