--- examples/wasm_custom_draw/src/main.rs
@@ -1,88 +1,0 @@
-use eframe::{App, CreationContext};
-use egui::Context;
-use egui_graphs::{DefaultEdgeShape, Graph, GraphView, SettingsInteraction, SettingsNavigation};
-use node::NodeShapeAnimated;
-use petgraph::{
-    stable_graph::{DefaultIx, StableGraph},
-    Directed,
-};
-
-mod node;
-
-pub struct CustomDrawApp {
-    g: Graph<(), (), Directed, DefaultIx, NodeShapeAnimated>,
-}
-
-impl CustomDrawApp {
-    fn new(_: &CreationContext<'_>) -> Self {
-        let g = generate_graph();
-        Self { g: Graph::from(&g) }
-    }
-}
-
-impl App for CustomDrawApp {
-    fn update(&mut self, ctx: &Context, _: &mut eframe::Frame) {
-        egui::CentralPanel::default().show(ctx, |ui| {
-            ui.add(
-                &mut GraphView::<_, _, _, _, NodeShapeAnimated, DefaultEdgeShape>::new(&mut self.g)
-                    .with_navigations(
-                        &SettingsNavigation::default()
-                            .with_fit_to_screen_enabled(false)
-                            .with_zoom_and_pan_enabled(true),
-                    )
-                    .with_interactions(
-                        &SettingsInteraction::default()
-                            .with_dragging_enabled(true)
-                            .with_node_selection_enabled(true)
-                            .with_edge_selection_enabled(true),
-                    ),
-            );
-        });
-    }
-}
-
-fn generate_graph() -> StableGraph<(), ()> {
-    let mut g = StableGraph::new();
-
-    let a = g.add_node(());
-    let b = g.add_node(());
-    let c = g.add_node(());
-
-    g.add_edge(a, b, ());
-    g.add_edge(b, c, ());
-    g.add_edge(c, a, ());
-
-    g
-}
-
-// When compiling natively:
-#[cfg(not(target_arch = "wasm32"))]
-fn main() {
-    let native_options = eframe::NativeOptions::default();
-    eframe::run_native(
-        "egui_graphs_custom_draw_demo",
-        native_options,
-        Box::new(|cc| Ok(Box::new(CustomDrawApp::new(cc)))),
-    )
-    .unwrap();
-}
-
-// When compiling to web using trunk:
-#[cfg(target_arch = "wasm32")]
-fn main() {
-    // Redirect `log` message to `console.log` and friends:
-    eframe::WebLogger::init(log::LevelFilter::Debug).ok();
-
-    let web_options = eframe::WebOptions::default();
-
-    wasm_bindgen_futures::spawn_local(async {
-        eframe::WebRunner::new()
-            .start(
-                "the_canvas_id", // hardcode it
-                web_options,
-                Box::new(|cc| Box::new(CustomDrawApp::new(cc))),
-            )
-            .await
-            .expect("failed to start eframe");
-    });
-}
